[Issue_95] To create an AssertionError, just invoke the constructor that takes Object
I also changed the other exception messages to emphasize that Docker wasn't available, with the 'we can't report this using the exception type you wanted' as a parenthetical point. This latter is kind of a nit as someone configuring a type that won't work is not likely to be confused by the existing message.